[level 1] Title: 두 정수 사이의 합, Time: 17.79 ms, Memory: 85.6 MB -BaekjoonHub

+++ 프로그래머스/1/12912. 두 정수 사이의 합/두 정수 사이의 합.java
@@ -10,7 +10,7 @@ class Solution {
         }
         
         for(int i=small; i<=big;i++)
-            sum += i;
+            sum += (long)i;
         
         return sum;
     }
